Refactor on helpers + now a welcome message is sent to new users that join a channel
