+++ helpers/mention_helper.js
@@ -1,25 +1,6 @@
 const {User} = require('../models/user');
 const botHelper = require('./bot_helper');
 
-
-function parseUrl(url) {
-  const protocol = 'https://';
-  const urlEnd = url.indexOf('/', protocol.length);
-  return url.substr(protocol.length, urlEnd - protocol.length);
-};
-
-function parsePath(url) {
-  const protocol = 'https://';
-  const path = url.substr(protocol.length);
-  const slashPos = path.indexOf('/');
-  return path.substr(slashPos);
-};
-
-function parseBotCommand(message, botName) {
-  return message.substr(message.indexOf(botName) + 1 + botName.length);
-}
-
-
 async function handleMentions(workspace, channel, message, sender) {
   const atPos = message.text.indexOf('@');
   if (atPos < 0) return;
@@ -28,17 +9,7 @@ async function handleMentions(workspace, channel, message, sender) {
   const user = await User.findOne({name: name});
   if (!user) return;
   if (user.url) { // El usuario es un bot si la url esta seteada
-    const hostname = parseUrl(user.url);
-    const botCommand = parseBotCommand(message.text, user.name);
-    const body = JSON.stringify({
-      sender: sender.email,
-      message: botCommand,
-      channel: channel.name,
-      workspace: workspace.name
-    });
-
-    const path = parsePath(user.url);
-    botHelper.sendRequest(hostname, path, body);
+    botHelper.sendRequest(workspace, channel, user, message, sender);
   }
 };
