+++ tests/integration/routes/channels.test.js
@@ -4,6 +4,7 @@ const {Workspace} = require('../../../models/workspace');
 const {Channel} = require('../../../models/channel');
 const {Page} = require('../../../models/page');
 const firebase = require('../../../helpers/firebase_helper');
+const botHelper = require('../../../helpers/bot_helper');
 
 let server;
 
@@ -64,6 +65,8 @@ describe('/api/channels', ()=> {
     firebase.subscribeToTopic = jest.fn();
     firebase.sendMessageToTopic = jest.fn();
     firebase.unsubscribeFromTopic = jest.fn();
+    botHelper.addTitoTo = jest.fn();
+    botHelper.sendWelcomeMessage = jest.fn();
   });
 
   afterAll(async ()=> {
@@ -118,13 +121,14 @@ describe('/api/channels', ()=> {
 
     it('should return new channel if request is valid', async ()=> {
       const response = await execute(token);
-
       expect(response.status).toBe(200);
       expect(Object.keys(response.body)).toEqual(
           expect.arrayContaining([
             '_id', 'name', 'welcomeMessage', 'description', 'isPrivate'
           ])
       );
+      expect(botHelper.addTitoTo).toHaveBeenCalled();
+      expect(botHelper.sendWelcomeMessage).toHaveBeenCalled();
     });
 
     it('should return 200 if user member creates a channel', async ()=> {
@@ -136,6 +140,8 @@ describe('/api/channels', ()=> {
             '_id', 'name', 'welcomeMessage', 'description', 'isPrivate'
           ])
       );
+      expect(botHelper.addTitoTo).toHaveBeenCalled();
+      expect(botHelper.sendWelcomeMessage).toHaveBeenCalled();
     });
 
     it('should return 400 if name is missing', async ()=> {
@@ -366,6 +372,7 @@ describe('/api/channels', ()=> {
       });
       users.push(userEmail);
       expect(usersEmails).toEqual(expect.arrayContaining(users));
+      expect(botHelper.sendWelcomeMessage).toHaveBeenCalled();
     });
 
     it(`should return 403 if user doesn't belong to the channel`, async () => {
